fix(issues): clear stale metadata on context switch, invalidate Linear cache
- Clear previous repo/team metadata before fetching new options to
  prevent cross-context items from appearing during fast switches
- Invalidate Linear issue cache entry (fetchedAt=0) in patchLinearIssue
  so fetchLinearIssue triggers a real IPC fetch after mutations
- Extract useImmediateMutation to its own file to stay under max-lines
